geolocation using redux setView action, layer actions needed for full implementation
